lib/build_OID_registry: fix reproducibility issues
The script build_OID_registry captures the full path of itself
in the generated data. This causes reproduciblity issues as the
path is captured and packaged.

We use the basename of the script instead, and that allows us
to be reprodicible, with slightly less information captured in
the output data (but the generating script can still easily
be found).

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Upstream-Status: Inappropriate

Taken from linux-yocto, v5.15/standard/base.

Signed-off-by: Anuj Mittal <anuj.mittal@intel.com>