QNX: rtc-cmos: Disable RTC alarm when ACPI does not report RTC irq
The virtual RTC provided by QNX does not support alarms. As a result qvm
never binds irq 8 to the virtual RTC and may use that irq for other
virtual devices. In such cases the kernel will complain that:

genirq: Flags mismatch irq 8. 00000000 (guest_shm) vs. 00000000 (rtc0)

That will prevent the real user of that irq from registering the handler
and cause the impacted virtual device to malfunction.

This patch makes the kernel skip RTC irq if it is not reported in the
ACPI namespace, so that we can provide a I/O-port-only RTC device object
to the kernel to make it aware that no irq is to be bound to the RTC.

Signed-off-by: Junjie Mao <junjie.mao@intel.com>
Signed-off-by: Yuanyuan Zhao <yuanyuan.zhao@intel.com>